Combine multiple words into a single string

When a value from an inline R expression is a character vector of multiple elements, we may want to combine them into a phrase like a and b, or a, b, and c. That is what this a helper function does.

Usage

combine_words(
  words,
  sep = ", ",
  and = " and ",
  before = "",
  after = before,
  oxford_comma = TRUE
)

Source

Adapted from knitr::combine_words() in the knitr package.

Arguments

words

A character vector.

sep

Separator to be inserted between words.

and

Character string to be prepended to the last word.

before, after

A character string to be added before/after each word.

oxford_comma

Whether to insert the separator between the last two elements in the list.

Value

A character string

Details

If the length of the input words is smaller than or equal to 1, words is returned. When words is of length 2, the first word and second word are combined using the and string, or if blank, sep if is used. When the length is greater than 2, sep is used to separate all words, and the and string is prepended to the last word.

Author

Yihui Xie xie@yihui.name (ORCID)

Examples

combine_words("a")
#> [1] "a"
combine_words(c("a", "b"))
#> [1] "a and b"
combine_words(c("a", "b", "c"))
#> [1] "a, b, and c"
combine_words(c("a", "b", "c"), sep = " / ", and = "")
#> [1] "a / b / c"
combine_words(c("a", "b", "c"), and = "")
#> [1] "a, b, c"
combine_words(c("a", "b", "c"), before = "\"", after = "\"")
#> [1] "\"a\", \"b\", and \"c\""
combine_words(c("a", "b", "c"), before = "\"", after = "\"", oxford_comma = FALSE)
#> [1] "\"a\", \"b\" and \"c\""