Apply a function to each element of a vector.

Apply a function to each element of a vector.

Usage

map_chr(.x, .f, ...)

Source

purr-like functions in staticimports package

Author

Winston Chang winston@stdout.org